requirements.txt
streamlit
easyocr
opencv-python
numpy
deep-translator
Pillow